requirements.txt
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0

matplotlib==3.8.3
seaborn==0.13.2

nibabel==5.2.1
nilearn==0.10.4

scikit-learn==1.4.1
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

tqdm==4.66.2
pyyaml==6.0.1

